Move the websocket wire into its own module
`wire.py` keeps the protocol every wire implements. `websocket_wire.py` holds the websocket wire and the two ends of a websocket session, beside `grpc_wire.py`.

Ticket: Positronic-Robotics/internal#1191 #refs
